Improve concurrency of the event store implementation (#109)
Using the go -race tool a couple of issues were discovered with the SubscribedCache. To solve this mutex locks are introduced and functionality is introduced to copy aggregate objects. This allows the updating of objects to be done in a new object, while the old version is still available in the cache to read.