Introduced Gonum-based graph operations (#111)
This PR adds two related packages. First this change introduces a Gonum-based graph implementation for improved performance and easier access to advanced, well-tested graph functionality. In a future PR the implicit graph implementation in the scheduler will be replaced with this one.

Partly using this new graphing support to detect circular dependencies, this PR also adds a new validation package. It aims to replace the existing ad-hoc implementation with a more structured, and consistent API. Likewise, in a future PR the current validation will be replaced with this validation package.

Specifically this change...

- Added gonum dependency
- Updated a couple of scripts (versions)
- Added new validation package in types package