Add Javascript and Repeat functions to internal fnenv (#110)
This change contains various changes to the fnenv/native package. It adds the javascript function, which allows users to create tasks using javascript. In the backend it is evaluated using the already in place Otto javascript engine. The idea behind it is that it will allow users to trivial functions and to provide more self-describing workflows (for demos or examples).

This also adds a prototype of the repeat function, which is in anticipation of the dynamic/inline workflow support. Although not usable in the current state, it shows how the dynamic workflows can be used to implement interesting functions.

Finally, the types/typedvalues package as part of the dynamic workflows milestone, undergoes some refactoring as well as adding initial support for the workflow type.

* Renamed a couple of things in the pubsub impl

* Add JavaScript and (initial) repeat functions to internal fnenv