Add invocationId to the TaskInvocation to avoid dangling tasks (#115)
This change adds an workflow invocation id to the TaskInvocationSpec. The idea behind this is that a task always runs within the scope of a workflow. Additionally, it is often needed to lookup or know a what invocation a task invocation belongs to. For these reasons it seems logical to include the workflow invocation id explicitely in the spec. Aside from that this PR centralizes and improves the validation of various models. All public API functions in the api package now have validation.