Improve workflow definition (#118)
This change aims to simplify the definition of workflows in Golang. The best example of this you can find in the new TestLongRunningWorkflowInvocation test. This test aims to cover the correct execution of parallel branches in a workflow. As the workflow definition in Golang is rather convenient, and to show that YAML doesn't necessarily have to be the only way to define a workflow, this change adds the protobuf parser, which, as it says, simply unmarshals the protobuf into a workflowspec. A metaparser is added to manage parsing files in using one or more of the parsers. This functionality is also added to the parse command in the wfcli.