Namespaced proto files (#153)
As part of a cleanup of stale, but as-good-as-finished, PRs - this PR adds namespacing to the protofiles, which makes our use of protobuf/grpc more compliant with best practices. The old setup caused issues with grpc-related tooling due to our.... peculiar setup. This PR fixes this by namespacing all the protofiles correctly.