Set content-length when setting new body (#156)
The Content-Length header in Golang http requests is only set when using NewRequest (instead of being set when the actual request is send). This means that when setting the Body field after the initial construction of the request, the content-length will remain unset unless explicitly set by the user.

To fix this issue, this PR explicitly sets the contentLength when setting the body.

As a flyby this PR also changes the default method to be used for fission requests to post, as this used to be (and still should be) the default.