Reduced addition of false positive missing evaluations
We used to add all evaluations that are not in the workQueue (invocationController) to the workQueue, even those that are currently being evaluated. This led to potentially a lot of unnecesary evaluations being added to workqueue. This change prevents these from being added, by checking whether the evaluations are locked (== being evaluated)