Replaced unbounded task execution in controller with executor
The executor and the corresponding task queue are responsible for executing the tasks received from the controller. By limiting the parallelism we ensure that the workflow engine does not simply break at >1000 parallel workflow invocations. The task queue also provides a way to provide a temporary buffer in times that the workflow engine is overloaded