Replaced aggregates with projectors
The aggregate models initially served as a way to do the updating of the entities (optionally) in-place. This was never used, mainly for to thread-safety, which made the aggregates solely a cumbersome layer in-between the application logic and the types. By moving the applying of events to explicit Projectors, the code can now use solely the core types.