Replaced controller implementations
The original controller implementation had become cluttered and unstructured over the last year. It also contained an initial prototype of a controller DSL (with rules-actions), which turned out to add more complexity than it solved. The new implementation removes most of this clutter in favor of a simple structure. In the new controllers there is an explicit controller struct for each invocation or workflow, which eliminates the need for the whole evalstates approach of the original implementation. The change also introduces the notion of the executor, which is the component responsible for executing the tasks generated by the controller. This improves the parallelism of the controller.