Fixed premature controller evaluation due to TaskAdded event
Due to the skipping of a TaskStarted event the arrival of the InvocationTaskAdded event would be just before the task completes. This means that it would trigger a controller evaluation before the state of the workflow was updated to reflect that the task was completed. The controller would then, based on the old state, reschedule the task again. We fixed this for now by simply ignoring the InovcationTaskAdded events, since it would never result in a useful evaluation anyway.